Rename dir for zod schema in schema request
